source/typely/usr/share/typely/requirements.txt
httpx>=0.28.0
faster-whisper>=1.1.0
huggingface-hub>=0.24.0
numpy>=1.26.0
PySide6>=6.7.0
python-xlib>=0.33
sounddevice>=0.4.7
webrtcvad-wheels>=2.0.14
